[internal] Consolidate PyO3 native extensions
