+++ src/python/pants/engine/internals/native_engine.pyi
@@ -8,6 +8,7 @@ from typing import Any, Sequence, TextIO, Tuple
 
 from typing_extensions import Protocol
 
+from pants.engine.fs import PathGlobs
 from pants.engine.internals.scheduler import Workunit, _PathGlobsAndRootCollection
 from pants.engine.internals.session import SessionValues
 from pants.engine.process import InteractiveProcessResult
@@ -15,6 +16,86 @@ from pants.engine.process import InteractiveProcessResult
 # TODO: black and flake8 disagree about the content of this file:
 #   see https://github.com/psf/black/issues/1548
 # flake8: noqa: E302
+
+# ------------------------------------------------------------------------------
+# Scheduler
+# ------------------------------------------------------------------------------
+
+class PyExecutor:
+    def __init__(self, core_threads: int, max_threads: int) -> None: ...
+
+# ------------------------------------------------------------------------------
+# FS
+# ------------------------------------------------------------------------------
+
+class PyDigest:
+    def __init__(self, fingerprint: str, serialized_bytes_length: int) -> None: ...
+    @property
+    def fingerprint(self) -> str: ...
+    @property
+    def serialized_bytes_length(self) -> int: ...
+    def __eq__(self, other: PyDigest | Any) -> bool: ...
+    def __hash__(self) -> int: ...
+    def __repr__(self) -> str: ...
+
+class PySnapshot:
+    def __init__(self) -> None: ...
+    @classmethod
+    def _create_for_testing(
+        cls, digest: PyDigest, files: Sequence[str], dirs: Sequence[str]
+    ) -> PySnapshot: ...
+    @property
+    def digest(self) -> PyDigest: ...
+    @property
+    def dirs(self) -> tuple[str, ...]: ...
+    @property
+    def files(self) -> tuple[str, ...]: ...
+    def __eq__(self, other: PySnapshot | Any) -> bool: ...
+    def __hash__(self) -> int: ...
+    def __repr__(self) -> str: ...
+
+def default_cache_path() -> str: ...
+
+# TODO: Really, `paths` should be `Sequence[str]`. Fix and update call sites so that we don't
+#  cast to `tuple()` when not necessary.
+def match_path_globs(path_globs: PathGlobs, paths: tuple[str, ...]) -> str: ...
+
+
+# ------------------------------------------------------------------------------
+# Workunits
+# ------------------------------------------------------------------------------
+
+def all_counter_names() -> list[str]: ...
+
+
+# ------------------------------------------------------------------------------
+# Nailgun
+# ------------------------------------------------------------------------------
+
+class PyNailgunClient:
+    def __init__(self, port: int, executor: PyExecutor) -> None: ...
+    def execute(self, command: str, args: list[str], env: dict[str, str]) -> int: ...
+
+class PantsdConnectionException(Exception):
+    pass
+
+class PantsdClientException(Exception):
+    pass
+
+# ------------------------------------------------------------------------------
+# Testutil
+# ------------------------------------------------------------------------------
+
+class PyStubCASBuilder:
+    def always_errors(self) -> PyStubCASBuilder: ...
+    def build(self, executor: PyExecutor) -> PyStubCAS: ...
+
+class PyStubCAS:
+    @classmethod
+    def builder(cls) -> PyStubCASBuilder: ...
+    @property
+    def address(self) -> str: ...
+
 
 class RawFdRunner(Protocol):
     def __call__(
@@ -141,32 +222,6 @@ def lease_files_in_graph(scheduler: PyScheduler, session: PySession) -> None: ..
 def strongly_connected_components(
     adjacency_lists: Sequence[Tuple[Any, Sequence[Any]]]
 ) -> Sequence[Sequence[Any]]: ...
-
-class PyDigest:
-    def __init__(self, fingerprint: str, serialized_bytes_length: int) -> None: ...
-    @property
-    def fingerprint(self) -> str: ...
-    @property
-    def serialized_bytes_length(self) -> int: ...
-    def __eq__(self, other: PyDigest | Any) -> bool: ...
-    def __hash__(self) -> int: ...
-    def __repr__(self) -> str: ...
-
-class PySnapshot:
-    def __init__(self) -> None: ...
-    @classmethod
-    def _create_for_testing(
-        cls, digest: PyDigest, files: Sequence[str], dirs: Sequence[str]
-    ) -> PySnapshot: ...
-    @property
-    def digest(self) -> PyDigest: ...
-    @property
-    def dirs(self) -> tuple[str, ...]: ...
-    @property
-    def files(self) -> tuple[str, ...]: ...
-    def __eq__(self, other: PySnapshot | Any) -> bool: ...
-    def __hash__(self) -> int: ...
-    def __repr__(self) -> str: ...
 
 class PyExecutionRequest:
     def __init__(
