+++ src/rust/engine/Cargo.toml
@@ -26,7 +26,6 @@ members = [
   "cache",
   "client",
   "concrete_time",
-  "engine_pyo3",
   "fs",
   "fs/brfs",
   "fs/fs_util",
@@ -68,7 +67,6 @@ default-members = [
   "cache",
   "client",
   "concrete_time",
-  "engine_pyo3",
   "fs",
   "fs/fs_util",
   "fs/store",
