--- src/rust/engine/engine_pyo3/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-version = "0.0.1"
-edition = "2018"
-name = "engine_pyo3"
-authors = [ "Pants Build <pantsbuild@gmail.com>" ]
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-
-[features]
-# NB: To actually load this crate from python, the `extension-module` feature must be enabled. But
-# unfortunately, enabling `extension-module` causes tests linked against `pyo3` to fail. We
-# define a feature to enable that, but we do not enable it by default: someone building this module
-# in order to extract `libengine_pyo3.so` should pass `cargo build .. --features=engine_pyo3/extension-module`.
-#  see https://github.com/PyO3/pyo3/issues/340
-extension-module = ["pyo3/extension-module"]
-default = []
-
-[dependencies]
-fs = { path = "../fs" }
-nailgun = { path = "../nailgun" }
-parking_lot = "0.11"
-# TODO: Switch back to official PyO3 once https://github.com/PyO3/pyo3/pull/1990 is released.
-pyo3 = { git = "https://github.com/davidhewitt/pyo3.git", rev = "51b6b947b82e61adad61524f9c58e3b1c89247a5" }
-task_executor = { path = "../task_executor" }
-testutil_mock = { package = "mock", path = "../testutil/mock" }
-workunit_store = { path = "../workunit_store" }
-
-[build-dependencies]
-pyo3-build-config = "0.15"
