--- src/rust/engine/engine_pyo3/src/externs/mod.rs
@@ -1,4 +1,0 @@
-// Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
-// Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-mod interface;
