+++ tests/python/pants_test/integration/remote_cache_integration_test.py
@@ -1,7 +1,7 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.engine.internals.native_engine_pyo3 import PyExecutor, PyStubCAS
+from pants.engine.internals.native_engine import PyExecutor, PyStubCAS
 from pants.option.global_options import RemoteCacheWarningsBehavior
 from pants.option.scope import GLOBAL_SCOPE_CONFIG_SECTION
 from pants.testutil.pants_integration_test import run_pants
